pull out the mirror code as it's not possiable on windows
